--- processor/src/main/resources/kotli/template/multiplatform/compose/platform/backend/title.md
@@ -1,1 +1,0 @@
-Backend
