--- processor/src/main/resources/kotli/template/multiplatform/compose/platform/backend/usage.md
@@ -1,5 +1,0 @@
-## Run
-
-```
-./gradlew :backend:run
-```
