+++ processor/src/main/resources/kotli/template/multiplatform/compose/platform/client/title.md
@@ -1,0 +1,1 @@
+Client
