+++ processor/src/main/resources/kotli/template/multiplatform/compose/platform/server/ktor/description.md
@@ -1,5 +1,5 @@
 Despite it not being common practice to keep frontend and backend code in the same repository due to differing project lifecycles and developer involvement, it is possible to organize the repository as a fully functional full-stack project.
 
-The backend target adds a pre-configured module to the project, which can serve as the backend for your app.
+The Ktor Server target adds a pre-configured module to the project, which can serve as the backend for your app.
 
 If you also include a JS target, the backend module will incorporate additional Gradle tasks to run and build the frontend and backend together, facilitating faster debugging of web-based solutions.
