+++ processor/src/main/resources/kotli/template/multiplatform/compose/platform/server/title.md
@@ -1,0 +1,1 @@
+Server
