--- processor/src/main/resources/kotli/template/multiplatform/compose/platform/title.md
@@ -1,1 +1,0 @@
-Platform
